bug fixed

+++ classfile/cp_invoke_dynamic.go
@@ -8,12 +8,12 @@ CONSTANT_MethodHandle_info {
 }
 */
 type ConstantMethodHandleInfo struct {
-  referenceKind   uint16
+  referenceKind   uint8
   referenceIndex  uint16
 }
 
 func (self *ConstantMethodHandleInfo) readInfo(reader *ClassReader) {
-  self.referenceKind = reader.readUint16()
+  self.referenceKind = reader.readUint8()
   self.referenceIndex = reader.readUint16()
 }
